+++ src_c/vprofile.c
@@ -59,8 +59,8 @@ static PyObject *voigt(PyObject *self, PyObject *args){
     return NULL;
 
   /* Get array sizes:                                                       */
-  nLor = PyArray_DIM(lorentz, 0);
-  nDop = PyArray_DIM(doppler, 0);
+  nLor = (int)PyArray_DIM(lorentz, 0);
+  nDop = (int)PyArray_DIM(doppler, 0);
 
   for   (m=0; m<nLor; m++){
     for (n=0; n<nDop; n++){
